refactor: improve cart modal store and extract mock data
- Update cart modal store to include products array in state
- Remove unused searchQuery and setCartQuery from store
- Extract mock cart items to separate file for better organization
- Simplify CartModal component props by removing unused type
- Add proper type safety for cart products in store
- Improve code organization and maintainability
