+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "concurrently \"next dev --turbopack\" \"json-server --watch db.json --port ${JSON_PORT:-9000}\"",
+    "dev": "concurrently --kill-others-on-fail \"next dev --turbopack\" \"json-server --watch db.json --port ${JSON_PORT:-9000}\"",
     "dev:next": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
