+++ src/widgets/header/ui/header-actions/header-actions.module.scss
@@ -1,3 +1,3 @@
 .container {
-
+display: flex;
 }
